Changed class FuiParameterView into a container of all possible parameter lists, such that we don't need to remove and add widgets when selecting functions of different types, or when switching function type.
Use Q3ScrollView for scrolling the parameter views, but
avoid the use of QGridLayout in the combination with this class.
There seems to be a problem with the columns not adjusting their width
when the main widget size changes. Instead, the row widgets are
positioned explicitly by the overridden resizeEvent() method.
This is also how it was done in the old FFuQtTableView class.