Simplify the FuiInputSelector widget by inheriting FFuComponentBase instead of FFuMultUIComponent, and let its Qt sub-class inherit FFuQtLabelFrame instead of having it as member. Use QLayout managers to position the child widgets.
Add container class FuiInputSelectors to be used for multi-argument
functions. It contains a map of widgets using the number of arguments
as key and where only one is displayed at any time depending on how
many arguments the function have. Each of these widgets then contain
a list of FuiInputSelector instances arranged in a scrollable list.
This replaces the use of FFuTable which did not work so well when
using layout managers for the positioning.